Date range selection and data semi-download automation

+++ Src/Data.py
@@ -1,18 +1,22 @@
 import pandas as pd
 import requests
 import os
+from datetime import datetime
+from Src.Settings import remove_all_files
 """"""""
 def get_data(currencies):
     for currency in currencies:
-        csv_url = f"https://stooq.pl/q/d/l/?s={currency}&i=d"  # Adres URL do pobrania pliku CSV
+        if not os.path.exists(os.getcwd() + f"/Exchange_rates/{currency}_{datetime.now().date()}.csv"):
+            remove_all_files(os.getcwd() + "/Exchange_rates",currency)
 
-        response = requests.get(csv_url)  # Wykonaj żądanie GET i pobierz zawartość pliku CSV
-        csv_content = response.content
+            csv_url = f"https://stooq.pl/q/d/l/?s={currency}&i=d"  # Adres URL do pobrania pliku CSV
+            response = requests.get(csv_url)  # Wykonaj żądanie GET i pobierz zawartość pliku CSV
+            csv_content = response.content
 
-        file_path = os.path.join(os.getcwd(), f"Exchange_rates/{currency}.csv")  # Zapisz zawartość pliku CSV w katalogu projektu
-        with open(file_path, "wb") as csv_file:
-            csv_file.write(csv_content)
-        print(f"The CSV file was successfully downloaded and saved as {currency}.csv")
+            file_path = os.path.join(os.getcwd(), f"Exchange_rates/{currency}_{datetime.now().date()}.csv")  # Zapisz zawartość pliku CSV w katalogu projektu
+            with open(file_path, "wb") as csv_file:
+                csv_file.write(csv_content)
+            print(f"The CSV file was successfully downloaded and saved as {currency}_{datetime.now().date()}.csv")
 def create_dataframe_list_from_csv(path):
     dataframe_list = []                             # zwracamy słownik postaci 'nokpln' : df itp...
 
@@ -28,7 +32,9 @@ def create_dataframe_list_from_csv(path):
                 'Zamkniecie': 'Close'
             }
             df.rename(columns=new_column_names, inplace=True)
-            word = {os.path.splitext(filename)[0] : df}
+            new_name = os.path.splitext(filename)[0]
+            new_name = new_name.split('_')[0]
+            word = {new_name : df}
             dataframe_list.append(word)
 
     return dataframe_list
